chore(devnet-sdk): move kurtosis artifact fs to devnet-sdk
This change makes devnet-sdk a leaf package as far as the rest of the
monorepo is concerned.

The artifact package was never really specific to kurtosis-devnet
anyway, as it contains a generic filesystem abstraction to access
kurtosis enclave artifacts. So we might as well add it to the rest of
kurtosis helper functions in devnet-sdk.

Going forward we'll likely want to split it into a separate module, so
that other repositories can import it without needing to import the
whole monorepo.
