+++ kurtosis-devnet/pkg/deploy/deploy.go
@@ -9,15 +9,11 @@ import (
 	"log"
 	"os"
 
+	ktfs "github.com/ethereum-optimism/optimism/devnet-sdk/kt/fs"
+	"github.com/ethereum-optimism/optimism/devnet-sdk/shell/env"
 	"github.com/ethereum-optimism/optimism/kurtosis-devnet/pkg/kurtosis"
 	"github.com/ethereum-optimism/optimism/kurtosis-devnet/pkg/kurtosis/api/engine"
-	"github.com/ethereum-optimism/optimism/kurtosis-devnet/pkg/kurtosis/sources/artifact"
 	"github.com/ethereum-optimism/optimism/kurtosis-devnet/pkg/kurtosis/sources/spec"
-)
-
-const (
-	DevnetEnvArtifactName = "devnet"
-	DevnetEnvArtifactPath = "env.json"
 )
 
 type EngineManager interface {
@@ -144,13 +140,13 @@ func (d *Deployer) deployEnvironment(ctx context.Context, r io.Reader) (*kurtosi
 		return nil, fmt.Errorf("error deploying kurtosis package: %w", err)
 	}
 
-	env, err := ktd.GetEnvironmentInfo(ctx, spec)
+	info, err := ktd.GetEnvironmentInfo(ctx, spec)
 	if err != nil {
 		return nil, fmt.Errorf("error getting environment info: %w", err)
 	}
 
 	// Upload the environment info to the enclave.
-	fs, err := artifact.NewEnclaveFS(ctx, d.enclave)
+	fs, err := ktfs.NewEnclaveFS(ctx, d.enclave)
 	if err != nil {
 		return nil, fmt.Errorf("error getting enclave fs: %w", err)
 	}
@@ -158,15 +154,15 @@ func (d *Deployer) deployEnvironment(ctx context.Context, r io.Reader) (*kurtosi
 	envBuf := bytes.NewBuffer(nil)
 	enc := json.NewEncoder(envBuf)
 	enc.SetIndent("", "  ")
-	if err := enc.Encode(env); err != nil {
+	if err := enc.Encode(info); err != nil {
 		return nil, fmt.Errorf("error encoding environment: %w", err)
 	}
 
-	if err := fs.PutArtifact(ctx, DevnetEnvArtifactName, artifact.NewArtifactFileReader(DevnetEnvArtifactPath, envBuf)); err != nil {
+	if err := fs.PutArtifact(ctx, env.KurtosisDevnetEnvArtifactName, ktfs.NewArtifactFileReader(env.KurtosisDevnetEnvArtifactPath, envBuf)); err != nil {
 		return nil, fmt.Errorf("error putting environment artifact: %w", err)
 	}
 
-	return env, nil
+	return info, nil
 }
 
 func (d *Deployer) renderTemplate(buildDir string, urlBuilder func(path ...string) string) (*bytes.Buffer, error) {
